+++ substra/sdk/client.py
@@ -797,7 +797,25 @@ class Client:
 
 
         """
-        return self._list(schemas.Type.SummaryTask, filters, order_by, ascending)
+        return self._list(schemas.Type.Task, filters, order_by, ascending)
+
+    @logit
+    def list_task_input_assets(self, key: str) -> List[models.InputAsset]:
+        """List input assets for a specific task, the returned object is described
+        in the [models.InputAsset](sdk_models.md#InputAsset) model"""
+        return self._backend.list_task_input_assets(key)
+
+    @logit
+    def list_task_output_assets(self, key: str) -> List[models.OutputAsset]:
+        """List output assets for a specific task, the returned object is described
+        in the [models.OutputAsset](sdk_models.md#OutputAsset) model"""
+        return self._backend.list_task_output_assets(key)
+
+    @logit
+    def get_task_output_asset(self, key: str, identifier: str) -> models.OutputAsset:
+        """Get an output asset for a specific task with a defined identifier, the returned object is described
+        in the [models.OutputAsset](sdk_models.md#OutputAsset) model"""
+        return self._backend.get_task_output_asset(key, identifier)
 
     @logit
     def list_organization(self, *args, **kwargs) -> List[models.Organization]:
@@ -943,8 +961,8 @@ class Client:
         Returns:
             pathlib.Path: Path of the downloaded model
         """
-        task = self._backend.get(schemas.Type.Task, task_key)
-        model = task.outputs[identifier].value
+        task_output = self._backend.get_task_output_asset(task_key, identifier)
+        model = task_output.asset
         return self.download_model(model.key, folder)
 
     @logit