+++ tests/datastore.py
@@ -48,7 +48,7 @@ DATASET = {
 }
 
 DATA_SAMPLE = {
-    "key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd",
+    "key": "7096db67-c175-4b45-9fb3-ad785eebb347",
     "owner": "MyOrg1MSP",
     "data_manager_keys": ["d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd"],
     "creation_date": "2021-08-24T13:36:07.428974249Z",
@@ -174,7 +174,7 @@ TRAINTASK = {
     "start_date": "2021-10-12T09:28:06.947765800Z",
     "end_date": "2021-10-12T09:30:04.705947400Z",
     "inputs": [
-        {"identifier": "opener", "asset_key": "a67b9387-fd80-429a-bc2f-034fac430b0f"},
+        {"identifier": "opener", "asset_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd"},
         {"identifier": "datasamples", "asset_key": "3180e12c-a821-434a-ad8a-a341076c7f98"},
         {"identifier": "datasamples", "asset_key": "21bb59ca-abd4-4154-b04a-44a92556a078"},
         {"identifier": "datasamples", "asset_key": "67512646-2464-4521-84de-419b1b307d30"},
@@ -187,23 +187,30 @@ TRAINTASK = {
                 "download": {"public": True, "authorized_ids": ["MyOrg1MSP"]},
             },
             "transient": False,
-            "value": {
-                "key": "6f0ee20a328044fb89e70ee5d219fa0c",
-                "compute_task_key": "30c283be-d385-424e-94a6-4d8538275260",
-                "address": {
-                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
-                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
-                },
-                "permissions": {
-                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                },
-                "owner": "MyOrg1MSP",
-                "creation_date": "2021-08-24T13:36:07.393646367Z",
-            },
         },
     },
 }
+
+TRAINTASK_MODEL = {
+    "asset": {
+        "key": "6f0ee20a328044fb89e70ee5d219fa0c",
+        "compute_task_key": "30c283be-d385-424e-94a6-4d8538275260",
+        "address": {
+            "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+            "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
+        },
+        "permissions": {
+            "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+            "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+        },
+        "owner": "MyOrg1MSP",
+        "creation_date": "2021-08-24T13:36:07.393646367Z",
+    },
+    "identifier": "model",
+    "kind": "ASSET_MODEL",
+}
+
+TRAINTASK_MODEL_RESPONSE = {"count": 1, "previous": None, "next": None, "results": [TRAINTASK_MODEL]}
 
 AGGREGATETASK = {
     "key": "06207faf-1785-4fa9-4220-99a50dcfe064",
@@ -246,23 +253,30 @@ AGGREGATETASK = {
                 "download": {"public": True, "authorized_ids": ["MyOrg1MSP"]},
             },
             "transient": False,
-            "value": {
-                "key": "5f0ee20a328044fb89e70ee5d219fa0b",
-                "compute_task_key": "06207faf-1785-4fa9-4220-99a50dcfe064",
-                "address": {
-                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
-                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
-                },
-                "permissions": {
-                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                },
-                "owner": "MyOrg1MSP",
-                "creation_date": "2021-08-24T13:36:07.393646367Z",
-            },
         },
     },
 }
+
+AGGREGATETASK_MODEL = {
+    "asset": {
+        "key": "5f0ee20a328044fb89e70ee5d219fa0b",
+        "compute_task_key": "06207faf-1785-4fa9-4220-99a50dcfe064",
+        "address": {
+            "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+            "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
+        },
+        "permissions": {
+            "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+            "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+        },
+        "owner": "MyOrg1MSP",
+        "creation_date": "2021-08-24T13:36:07.393646367Z",
+    },
+    "identifier": "model",
+    "kind": "ASSET_MODEL",
+}
+
+AGGREGATETASK_MODEL_RESPONSE = {"count": 1, "previous": None, "next": None, "results": [AGGREGATETASK_MODEL]}
 
 COMPOSITE_TRAINTASK = {
     "key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
@@ -316,20 +330,6 @@ COMPOSITE_TRAINTASK = {
                 "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
             },
             "transient": False,
-            "value": {
-                "key": "dc9d114d1ed54936b57dd7fef1c0cbf4",
-                "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
-                "address": {
-                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
-                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
-                },
-                "permissions": {
-                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                },
-                "owner": "MyOrg1MSP",
-                "creation_date": "2021-08-24T13:36:07.393646367Z",
-            },
         },
         "local": {
             "permissions": {
@@ -337,22 +337,57 @@ COMPOSITE_TRAINTASK = {
                 "download": {"public": False, "authorized_ids": ["MyOrg2MSP"]},
             },
             "transient": False,
-            "value": {
-                "key": "ac4a1a82ced4419b83f4bba78c024b94",
-                "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
-                "address": {
-                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
-                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
-                },
-                "permissions": {
-                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
-                },
-                "owner": "MyOrg1MSP",
-                "creation_date": "2021-08-24T13:36:07.393646367Z",
-            },
         },
     },
+}
+
+COMPOSITE_TRAINTASK_SHARED = {
+    "asset": {
+        "key": "dc9d114d1ed54936b57dd7fef1c0cbf4",
+        "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
+        "address": {
+            "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+            "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
+        },
+        "permissions": {
+            "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+            "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+        },
+        "owner": "MyOrg1MSP",
+        "creation_date": "2021-08-24T13:36:07.393646367Z",
+    },
+    "identifier": "shared",
+    "kind": "ASSET_MODEL",
+}
+
+
+COMPOSITE_TRAINTASK_SHARED_RESPONSE = {"count": 1, "previous": None, "next": None, "results": [AGGREGATETASK_MODEL]}
+
+COMPOSITE_TRAINTASK_LOCAL = {
+    "asset": {
+        "key": "dc9d114d1ed54936b57dd7fef1c0cbf4",
+        "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
+        "address": {
+            "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+            "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
+        },
+        "permissions": {
+            "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+            "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+        },
+        "owner": "MyOrg1MSP",
+        "creation_date": "2021-08-24T13:36:07.393646367Z",
+    },
+    "identifier": "local",
+    "kind": "ASSET_MODEL",
+}
+
+
+COMPOSITE_TRAINTASK_LOCAL_RESPONSE = {
+    "count": 1,
+    "previous": None,
+    "next": None,
+    "results": [COMPOSITE_TRAINTASK_LOCAL],
 }
 
 COMPOSITE_TRAINTASK_DOING = {
@@ -455,7 +490,7 @@ PREDICTTASK = {
     "start_date": "2021-10-12T09:28:06.947765800Z",
     "end_date": "2021-10-12T09:30:04.705947400Z",
     "inputs": [
-        {"identifier": "opener", "asset_key": "a67b9387-fd80-429a-bc2f-034fac430b0f"},
+        {"identifier": "opener", "asset_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd"},
         {"identifier": "datasamples", "asset_key": "3180e12c-a821-434a-ad8a-a341076c7f98"},
         {"identifier": "datasamples", "asset_key": "21bb59ca-abd4-4154-b04a-44a92556a078"},
         {
@@ -513,7 +548,7 @@ TESTTASK = {
     "start_date": "2021-10-12T09:28:06.947765800Z",
     "end_date": "2021-10-12T09:30:04.705947400Z",
     "inputs": [
-        {"identifier": "opener", "asset_key": "a67b9387-fd80-429a-bc2f-034fac430b0f"},
+        {"identifier": "opener", "asset_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd"},
         {"identifier": "datasamples", "asset_key": "3180e12c-a821-434a-ad8a-a341076c7f98"},
         {"identifier": "datasamples", "asset_key": "21bb59ca-abd4-4154-b04a-44a92556a078"},
         {
@@ -533,10 +568,8 @@ TESTTASK = {
     },
 }
 
-TASK_LIST = [
-    {key: value for key, value in task.items() if key not in {"inputs", "outputs"}}
-    for task in [PREDICTTASK, AGGREGATETASK, COMPOSITE_TRAINTASK, TRAINTASK, TESTTASK]
-]
+TASK_LIST = [PREDICTTASK, AGGREGATETASK, COMPOSITE_TRAINTASK, TRAINTASK, TESTTASK]
+
 
 COMPUTE_PLAN = {
     "key": "e983a185-5368-bd0a-0190-183af9a8e560",
@@ -567,7 +600,7 @@ COMPUTE_PLAN_PERF = {
         {
             "compute_task": {
                 "key": "afa9c7b1-21b5-4e58-a098-79de4ecede35",
-                "data_manager_key": "a67b9387-fd80-429a-bc2f-034fac430b0f",
+                "data_manager_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd",
                 "function_key": "17f98afc-2b82-4ce9-b232-1a471633d020",
                 "rank": 1,
                 "epoch": None,
@@ -584,7 +617,7 @@ COMPUTE_PLAN_PERF = {
         {
             "compute_task": {
                 "key": "afa9c7b1-21b5-4e58-a098-79de4ecede35",
-                "data_manager_key": "a67b9387-fd80-429a-bc2f-034fac430b0f",
+                "data_manager_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd",
                 "function_key": "17f98afc-2b82-4ce9-b232-1a471633d020",
                 "rank": 2,
                 "epoch": None,
@@ -601,7 +634,7 @@ COMPUTE_PLAN_PERF = {
         {
             "compute_task": {
                 "key": "afa9c7b1-21b5-4e58-a098-79de4ecede35",
-                "data_manager_key": "a67b9387-fd80-429a-bc2f-034fac430b0f",
+                "data_manager_key": "d7dc4a5b-e81f-4f3f-b94c-5e2bbaca15cd",
                 "function_key": "17f98afc-2b82-4ce9-b232-1a471633d020",
                 "rank": 3,
                 "epoch": None,